requirements.txt
langchain
google-cloud-aiplatform
bs4
elasticsearch
flask
unstructured
selenium
faiss-cpu
openai
Chroma
chromadb
tiktoken
fastapi
uvicorn
playwright
selectolax